Changed coordinates to float in MapElement table, TDG and test

+++ Exposeum/Exposeum/Tables/MapElements.cs
@@ -5,8 +5,8 @@ namespace Exposeum.Tables
     {
         [PrimaryKey, AutoIncrement, Column("ID")]
         public int ID { get; set; }
-        public double uCoordinate { get; set; }
-        public double vCoordinate { get; set; }
+        public float uCoordinate { get; set; }
+        public float vCoordinate { get; set; }
         public int iconId { get; set; }
         public string discriminator { get; set; }
         public int visited { get; set; }
